+++ splitshifts-app/app/(logged-in)/change-password/page.tsx
@@ -1,4 +1,4 @@
-import ChangePasswordForm from "./components/ChangePasswordForm";
+import ChangePasswordForm from "./components/change-password-form";
 
 export default function ChangePasswordPage() {
   return (
