+++ splitshifts-app/app/(public)/(auth)/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from './components/SignUpForm';
+import SignUpForm from './components/signup-form';
 
 export default function SignUpPage() {
   return (
